Fix import/flake issues
