+++ canopen_monitor/parse/__init__.py
@@ -2,9 +2,8 @@
 for parsing CANOpen messages according to Object Definiton files or Electronic
 Data Sheet files, provided by the end user.
 """
-from .eds import EDS, load_eds_file, load_eds_files
+from .eds import EDS, load_eds_file, load_eds_files, DataType
 from .canopen import CANOpenParser
-from .utilities import DataType
 
 __all__ = [
     'CANOpenParser',
